app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.27.2
pydantic==2.9.2
python-multipart==0.0.12
python-dotenv==1.0.1
streamlit==1.41.1
requests==2.32.3
pytest>=8.0.0
pytest-asyncio>=0.23.0
respx>=0.21.0

